clean up models, move them to filmeddomain
