+++ filmeddata/build.gradle
@@ -1,4 +1,3 @@
 apply plugin: 'com.android.library'
 
 apply from: '../gradle/common.gradle'
-
