--- filmeddomain/src/main/java/me/joshshin/filmeddomain/models/FilmedMoviesResponse.kt
@@ -1,12 +1,0 @@
-package me.joshshin.filmeddomain.models
-
-/**
- * Created by Josh Shin on 4/22/18
- */
-
-interface FilmedMoviesResponse {
-    val page: Int?
-    val totalResults: Int?
-    val totalPages: Int?
-    val results: List<FilmedMovie>?
-}
